refactor:  `peaks ` `ModelSystem` in separate file
