+++ ProcessOptimizer/model_systems/__init__.py
@@ -9,11 +9,11 @@ from .noise_models import (
     parse_noise_model,
     noise_model_factory,
 )
-from .benchmarks import peaks_model_system as peaks
 from .gold_map import gold_map
 from .hart3 import hart3, hart3_no_noise
 from .hart6 import hart6, hart6_no_noise
 from .poly2 import poly2, poly2_no_noise
+from .peaks import peaks, peaks_no_noise
 
 __all__ = [
     "branin",
@@ -33,5 +33,6 @@ __all__ = [
     "poly2",
     "poly2_no_noise",
     "peaks",
+    "peaks_no_noise",
     "gold_map",
 ]
